requirements.txt
click>=8.1.3
dash>=2.18.2
dash-bootstrap-components>=1.0.0
flask>=3.0
flask-socketio>=5.0.0
graphviz>=0.20
lark>=1.1.5
matplotlib<3.10
networkx>=2.8.8
numpy>=1.23.0
plotly>=5.18
psutil>=5.9.0
pytest>=7.0.0
torch>=1.10.0
pygithub>=1.59
pyyaml>=6.0.1
selenium>=4.0
optuna>=3.0
fastapi>=0.68
dash>=2.0
python-dotenv>=1.0
webdriver-manager
tensorflow>=2.6
huggingface_hub>=0.16
transformers>=4.30
torchvision>=0.15
multiprocess>=0.70
tweepy==4.15.0
radon>=5.0
pandas>=1.3
scikit-learn>=1.0
scipy>=1.7
seaborn>=0.11
statsmodels>=0.13
sympy>=1.9
onnx>=1.10
onnxruntime>=1.10
flask_httpauth>=4.4
flask_cors>=3.1
